Fix size_t not always defined.
<vector> does not provide its definition according to the standard.
However we don't actually need size_t because the number of tasks we
deal with is largely inferior to even 32-bit integers.

+++ engine/buffered_task_scheduler.h
@@ -23,11 +23,11 @@ public:
 		_io_tasks.push_back(task);
 	}
 
-	inline size_t get_main_count() const {
+	inline unsigned int get_main_count() const {
 		return _main_tasks.size();
 	}
 
-	inline size_t get_io_count() const {
+	inline unsigned int get_io_count() const {
 		return _io_tasks.size();
 	}
